+++ citywok_ms/employee/routes.py
@@ -4,7 +4,17 @@ from citywok_ms.employee.forms import EmployeeForm
 from citywok_ms.employee.models import Employee
 from citywok_ms.file.forms import FileForm
 from citywok_ms.file.models import EmployeeFile, File
-from flask import Blueprint, current_app, flash, redirect, render_template, url_for
+from citywok_ms.task import compress_file
+from flask import (
+    Blueprint,
+    current_app,
+    flash,
+    redirect,
+    render_template,
+    url_for,
+    request,
+    send_file,
+)
 from flask_babel import _
 
 employee_bp = Blueprint("employee", __name__, url_prefix="/employee")
@@ -13,11 +23,26 @@ employee_bp = Blueprint("employee", __name__, url_prefix="/employee")
 @employee_bp.route("/")
 @visitor.require(401)
 def index():
+    keys = (
+        ("id", _("ID")),
+        ("first_name", _("First Name")),
+        ("last_name", _("Last Name")),
+        ("zh_name", _("Chinese Name")),
+        ("accountant_id", _("Accountant ID")),
+        ("sex", _("Sex")),
+        ("nif", _("NIF")),
+        ("niss", _("NISS")),
+    )
+    sort = request.args.get("sort") or "id"
+    desc = request.args.get("desc") or False
     return render_template(
         "employee/index.html",
         title=_("Employees"),
-        active_employees=Employee.get_active(),
-        suspended_employees=Employee.get_suspended(),
+        active_employees=Employee.get_active(sort, desc),
+        suspended_employees=Employee.get_suspended(sort, desc),
+        keys=keys,
+        sort=sort,
+        desc=desc,
     )
 
 
@@ -112,6 +137,8 @@ def upload(employee_id):
         )
         db.session.commit()
         current_app.logger.info(f"Upload employee file {db_file}")
+        compress_file.queue(db_file.id)
+
     elif file is not None:
         flash(
             _('Invalid file format "%(format)s".', format=File.split_file_format(file)),
@@ -120,3 +147,16 @@ def upload(employee_id):
     else:
         flash(_("No file has been uploaded."), "danger")
     return redirect(url_for("employee.detail", employee_id=employee_id))
+
+
+@employee_bp.route("/export/<export_format>")
+@manager.require(403)
+def export(export_format):
+    if export_format == "csv":
+        return send_file(
+            Employee.export_to_csv(), cache_timeout=0, download_name="Employees.csv"
+        )
+    elif export_format == "excel":
+        return send_file(
+            Employee.export_to_excel(), cache_timeout=0, download_name="Employees.xlsx"
+        )