+++ citywok_ms/supplier/routes.py
@@ -1,3 +1,4 @@
+from citywok_ms.task import compress_file
 from flask.globals import current_app
 from citywok_ms import db
 from citywok_ms.auth.permissions import manager, shareholder, visitor
@@ -5,7 +6,15 @@ from citywok_ms.file.forms import FileForm
 from citywok_ms.file.models import File, SupplierFile
 from citywok_ms.supplier.forms import SupplierForm
 from citywok_ms.supplier.models import Supplier
-from flask import Blueprint, flash, redirect, render_template, url_for
+from flask import (
+    Blueprint,
+    flash,
+    redirect,
+    render_template,
+    url_for,
+    request,
+    send_file,
+)
 from flask_babel import _
 
 supplier_bp = Blueprint("supplier", __name__, url_prefix="/supplier")
@@ -14,10 +23,25 @@ supplier_bp = Blueprint("supplier", __name__, url_prefix="/supplier")
 @supplier_bp.route("/")
 @visitor.require(401)
 def index():
+    keys = (
+        ("id", _("ID")),
+        ("name", _("Company Name")),
+        ("principal", _("Principal")),
+        ("abbreviation", _("Abbreviation")),
+        ("nif", _("NIF")),
+        ("iban", _("IBAN")),
+        ("contact", _("Contact")),
+        ("email", _("E-mail")),
+    )
+    sort = request.args.get("sort") or "id"
+    desc = request.args.get("desc") or False
     return render_template(
         "supplier/index.html",
         title=_("Suppliers"),
-        suppliers=Supplier.get_all(),
+        suppliers=Supplier.get_all(sort, desc),
+        keys=keys,
+        sort=sort,
+        desc=desc,
     )
 
 
@@ -82,6 +106,8 @@ def upload(supplier_id):
         )
         db.session.commit()
         current_app.logger.info(f"Upload supplier file {db_file}")
+        compress_file.queue(db_file.id)
+
     elif file is not None:
         flash(
             _('Invalid file format "%(format)s".', format=File.split_file_format(file)),
@@ -90,3 +116,16 @@ def upload(supplier_id):
     else:
         flash(_("No file has been uploaded."), "danger")
     return redirect(url_for("supplier.detail", supplier_id=supplier_id))
+
+
+@supplier_bp.route("/export/<export_format>")
+@manager.require(403)
+def export(export_format):
+    if export_format == "csv":
+        return send_file(
+            Supplier.export_to_csv(), cache_timeout=0, download_name="Suppliers.csv"
+        )
+    elif export_format == "excel":
+        return send_file(
+            Supplier.export_to_excel(), cache_timeout=0, download_name="Suppliers.xlsx"
+        )