+++ citywok_ms/utils/__init__.py
@@ -1,14 +1,18 @@
-SEX = [('M', 'Male'),
-       ('F', 'Female')]
+from flask_babel import lazy_gettext as _l
 
-ID = [('passport', 'Passport'),
-      ('id_card', 'ID Card'),
-      ('green_card', 'Permanent Resident Card'),
-      ('residence_permit', 'Residence Permit'),
-      ('other', 'Other')]
+SEX = [("M", _l("Male")), ("F", _l("Female"))]
+
+ID = [
+    ("passport", _l("Passport")),
+    ("id_card", _l("ID Card")),
+    ("green_card", _l("Permanent Resident Card")),
+    ("residence_permit", _l("Residence Permit")),
+    ("other", _l("Other")),
+]
 
 FILEALLOWED = tuple(
-    '''txt jpg jpe jpeg png gif svg bmp rtf
+    """txt jpg jpe jpeg png gif svg bmp rtf
     odf ods gnumeric abw doc docx xls xlsx
     csv ini json plist xml yaml yml gz bz2
-    zip tar tgz txz 7z pdf'''.split())
+    zip tar tgz txz 7z pdf""".split()
+)