+++ tests/test_employee.py
@@ -10,6 +10,7 @@ from citywok_ms.employee.models import Employee
 from citywok_ms.file.models import EmployeeFile
 from flask import request, url_for
 from wtforms.fields.simple import HiddenField, SubmitField
+import pandas as pd
 
 
 @pytest.mark.role("admin")
@@ -32,7 +33,7 @@ def test_index_get(client, user):
 
 @pytest.mark.role("admin")
 def test_index_get_with_employee(client, user, employee):
-    response = client.get(url_for("employee.index"))
+    response = client.get(url_for("employee.index", desc=True))
     data = response.data.decode()
 
     # state code
@@ -125,6 +126,7 @@ def test_new_post_invalid(client, user, employee):
     request_data = {
         "first_name": "NEW",
         "sex": "F",
+        "accountant_id": 1,
         "id_type": "passport",
         "id_number": "1",
         "id_validity": "2000-01-01",
@@ -150,6 +152,7 @@ def test_new_post_invalid(client, user, employee):
     assert "This NIF already existe" in data
     assert "This NISS already existe" in data
     assert "This IBAN already existe" in data
+    assert "This Accountant ID already existe" in data
 
     # database data
     assert db.session.query(Employee).count() == 2  # 2 employee created in fixture
@@ -349,9 +352,9 @@ def test_upload_get(client, user, employee, id):
 
 @pytest.mark.role("admin")
 @pytest.mark.parametrize("id", [1, 2])
-def test_upload_post_valid(client, user, employee, id):
+def test_upload_post_valid(client, user, employee, id, image):
     request_data = {
-        "file": (io.BytesIO(b"test"), "test.jpg"),
+        "file": (image, "test.jpg"),
     }
     response = client.post(
         url_for("employee.upload", employee_id=id),
@@ -402,3 +405,35 @@ def test_upload_post_invalid_empty(client, user, employee, id):
     assert response.status_code == 200
     assert request.url.endswith(url_for("employee.detail", employee_id=id))
     assert "No file has been uploaded." in html.unescape(data)
+
+
+@pytest.mark.role("admin")
+def test_export_csv(client, user, employee):
+    response = client.get(
+        url_for("employee.export", export_format="csv"),
+        follow_redirects=True,
+    )
+    data = response.data.decode()
+    assert response.status_code == 200
+    for name in Employee.columns_name.values():
+        assert str(name) in data
+    for employee in Employee.get_all(sort="id", desc=False):
+        for attr in Employee.__table__.columns:
+            assert str(getattr(employee, attr.name) or "-") in data
+
+
+@pytest.mark.role("admin")
+def test_export_excel(client, user, employee):
+    response = client.get(
+        url_for("employee.export", export_format="excel"),
+        follow_redirects=True,
+    )
+    data = pd.read_excel(response.data)
+    assert response.status_code == 200
+    for name in Employee.columns_name.values():
+        assert str(name) in data
+
+    for employee in Employee.get_all(sort="id", desc=False):
+        for attr in ("first_name", "last_name", "id"):
+            value = getattr(employee, attr)
+            assert value in data.values